refactor(runExceptionTest): 유닛 테스트의 테스트 케이스를 돌리는 runExceptionTest를 재사용성 높게 리펙터링
