+++ __tests__/moneyValidation.test.js
@@ -1,12 +1,7 @@
-import { Console } from '@woowacourse/mission-utils';
 import { validateMoney } from '../src/Validation/validateMoney.js';
 import { ERROR_MESSAGES } from '../src/Error.js';
+import { runExceptionTest } from '../src/Test/Test.js';
 
-const getLogSpy = () => {
-  const logSpy = jest.spyOn(Console, 'print');
-  logSpy.mockClear();
-  return logSpy;
-};
 const validateMoneyTestCases = [
   {
     description: '1000을 입력하면 올바르게 1000을 반환해야 한다.',
@@ -89,12 +84,4 @@ const validateMoneyTestCases = [
   },
 ];
 
-const runExceptionTest = async ({ description, input, expected, log }) => {
-  test(description, async () => {
-    const logSpy = getLogSpy();
-    const result = validateMoney(input);
-    if (expected) expect(result).toBe(expected);
-    if (log) expect(logSpy).toHaveBeenCalledWith(expect.stringContaining(log));
-  });
-};
-validateMoneyTestCases.forEach(runExceptionTest);
+validateMoneyTestCases.forEach(runExceptionTest(validateMoney));
